OAUTH2-251 Remove SB dependencies from the ScopeFinder from the module oauth2-provider-shortcut to ensure it is available during upgrade
