+++ modules/apps/oauth2-provider/oauth2-provider-shortcut/src/main/java/com/liferay/oauth2/provider/shortcut/internal/instance/lifecycle/OAuth2ProviderShortcutPortalInstanceLifecycleListener.java
@@ -18,11 +18,6 @@ import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.oauth2.provider.constants.ClientProfile;
 import com.liferay.oauth2.provider.constants.GrantType;
 import com.liferay.oauth2.provider.model.OAuth2Application;
-import com.liferay.oauth2.provider.scope.spi.application.descriptor.ApplicationDescriptor;
-import com.liferay.oauth2.provider.scope.spi.prefix.handler.PrefixHandler;
-import com.liferay.oauth2.provider.scope.spi.prefix.handler.PrefixHandlerFactory;
-import com.liferay.oauth2.provider.scope.spi.scope.finder.ScopeFinder;
-import com.liferay.oauth2.provider.scope.spi.scope.mapper.ScopeMapper;
 import com.liferay.oauth2.provider.service.OAuth2ApplicationLocalService;
 import com.liferay.oauth2.provider.util.OAuth2SecureRandomGenerator;
 import com.liferay.petra.string.StringBundler;
@@ -46,7 +41,6 @@ import com.liferay.portal.kernel.service.UserGroupService;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.service.UserService;
 import com.liferay.portal.kernel.util.AggregateResourceBundleLoader;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleLoader;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
@@ -59,13 +53,10 @@ import java.io.InputStream;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Set;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -82,44 +73,10 @@ import org.osgi.service.component.annotations.Reference;
 		"osgi.jaxrs.name=liferay-json-web-services-analytics",
 		"sap.scope.finder=true"
 	},
-	service = {
-		ApplicationDescriptor.class, PortalInstanceLifecycleListener.class,
-		PrefixHandlerFactory.class, ScopeFinder.class, ScopeMapper.class
-	}
+	service = PortalInstanceLifecycleListener.class
 )
 public class OAuth2ProviderShortcutPortalInstanceLifecycleListener
-	extends BasePortalInstanceLifecycleListener
-	implements ApplicationDescriptor, PrefixHandlerFactory, ScopeFinder,
-			   ScopeMapper {
-
-	@Override
-	public PrefixHandler create(
-		Function<String, Object> propertyAccessorFunction) {
-
-		return PrefixHandler.PASS_THROUGH_PREFIX_HANDLER;
-	}
-
-	@Override
-	public String describeApplication(Locale locale) {
-		return GetterUtil.getString(
-			ResourceBundleUtil.getString(
-				ResourceBundleUtil.getBundle(
-					locale,
-					OAuth2ProviderShortcutPortalInstanceLifecycleListener.
-						class),
-				"liferay-json-web-services-analytics-name"),
-			"liferay-json-web-services-analytics-name");
-	}
-
-	@Override
-	public Collection<String> findScopes() {
-		return _scopeAliasesList;
-	}
-
-	@Override
-	public Set<String> map(String scope) {
-		return Collections.singleton(scope);
-	}
+	extends BasePortalInstanceLifecycleListener {
 
 	@Override
 	public void portalInstanceRegistered(Company company) throws Exception {
